% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_conda.R
\name{conda_create_weartime}
\alias{conda_create_weartime}
\alias{use_weartime_condaenv}
\alias{unset_reticulate_python}
\alias{have_weartime_condaenv}
\title{Create Conda Environmment for \code{weartime}}
\usage{
conda_create_weartime(envname = "weartime", ...)

use_weartime_condaenv(envname = "weartime")

unset_reticulate_python()

have_weartime_condaenv(envname = "weartime")
}
\arguments{
\item{envname}{conda environment name for the modules to be installed}

\item{...}{additional arguments to pass to \code{\link[reticulate:conda-tools]{reticulate::conda_create()}}}
}
\value{
Nothing, called for side effects
}
\description{
Create Conda Environmment for \code{weartime}
}
