% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cnn_model.R
\name{download_cnn_model}
\alias{download_cnn_model}
\title{Download CNN for Wear Time Estimation}
\usage{
download_cnn_model(episode_window_sec = 7, outdir = NULL, verbose = TRUE)
}
\arguments{
\item{episode_window_sec}{define window length to create input features for the CNN model}

\item{outdir}{output directory for \code{h5} file}

\item{verbose}{print diagnostic messages}
}
\value{
A path to the \code{h5} file
}
\description{
Download CNN for Wear Time Estimation
}
\examples{
download_cnn_model(outdir = tempdir())
}
