% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_baseline.R
\name{wt_baseline}
\alias{wt_baseline}
\alias{wt_xyz}
\alias{wt_vmu}
\title{Standard Deviation Non-Wear}
\usage{
wt_baseline(
  accdata,
  sample_rate = NULL,
  std_threshold = 0.004,
  min_interval = 90L,
  use_vmu = FALSE,
  verbose = TRUE
)

wt_xyz(
  accdata,
  sample_rate = NULL,
  std_threshold = 0.004,
  min_interval = 90L,
  use_vmu = FALSE,
  verbose = TRUE
)

wt_vmu(..., min_interval = 105L, use_vmu = TRUE)
}
\arguments{
\item{accdata}{activity data, usually output from \code{\link{py_read_gt3x}},
and then imputed}

\item{sample_rate}{sample rate (integer) of the sampling frequency in Hertz from the header}

\item{std_threshold}{standard deviation threshold in g}

\item{min_interval}{minimum interval (in minutes) that determines non-wear}

\item{use_vmu}{Use vector magnitude}

\item{verbose}{print diagnostic messages}

\item{...}{not used}
}
\value{
A tibble of time and indicator of wear
}
\description{
Calculate non-wear time from raw acceleration data by finding intervals in which the acceleration standard deviation is below a std_threshold value
}
\note{
See \url{https://doi.org/10.1101/2020.07.08.20148015}
}
\examples{
path = system.file("extdata", "TAS1H30182785_2019-09-17.gt3x",
package = "weartime")
df = read.gt3x::read.gt3x(path, asDataFrame = TRUE,
                           imputeZeroes = TRUE,
                           verbose = TRUE)
out = wt_baseline(df)
}
