% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_hees.R
\name{wt_hees_2013}
\alias{wt_hees_2013}
\alias{wt_hees_2011}
\alias{wt_hees_optimized}
\title{Estimation of non-wear time periods based on Hees 2013 paper}
\usage{
wt_hees_2013(
  accdata,
  sample_rate = NULL,
  min_non_wear_time_window = 60L,
  window_overlap = 15L,
  std_threshold = 3,
  std_min_num_axes = 2L,
  value_range_mg_threshold = 50,
  value_range_min_num_axes = 2L,
  verbose = TRUE
)

wt_hees_2011(..., min_non_wear_time_window = 30L)

wt_hees_optimized(
  accdata,
  sample_rate = NULL,
  min_non_wear_time_window = 135L,
  window_overlap = 15L,
  std_threshold = 7,
  std_min_num_axes = 1L,
  value_range_mg_threshold = 1,
  value_range_min_num_axes = 1L,
  verbose = TRUE
)
}
\arguments{
\item{accdata}{activity data, usually output from \code{\link{py_read_gt3x}},
and then imputed}

\item{sample_rate}{sample rate (integer) of the sampling frequency in Hertz from the header}

\item{min_non_wear_time_window}{int (optional) minimum window length in minutes to be classified as non-wear time}

\item{window_overlap}{int (optional) basically the sliding window that progresses over the acceleration data. Defaults to 15 minutes.}

\item{std_threshold}{float (optional) standard deviation threshold in mg. Acceleration axes values below or equal this threshold can be considered non-wear time. Defaults to 3.0mg. Note that within the code we convert mg to g.}

\item{std_min_num_axes}{int (optional)  minimum number of axes used to check if acceleration values are below the \code{std_threshold} value. Defaults to 2 axes; meaning that at least 2  axes need to have values below a threshold value to be considered non wear time}

\item{value_range_mg_threshold}{float (optional) value range threshold value in mg. If the range of values within a window is below this threshold (meaning that there is very little change in acceleration over time) then this can be considered non wear time. Default to 50 mg. Note that within the code we convert mg to g}

\item{value_range_min_num_axes}{int (optional) minimum number of axes used to check if acceleration values range are below the value_range_mg_threshold value. Defaults to 2 axes; meaning that at least 2 axes need to have a value range below a threshold value to be considered non wear time}

\item{verbose}{print diagnostic messages}

\item{...}{not used}
}
\value{
A tibble of time and indicator of wear
}
\description{
Estimation of non-wear time periods based on Hees 2013 paper
}
\examples{
path = system.file("extdata", "TAS1H30182785_2019-09-17.gt3x",
package = "pygt3x")
res = pygt3x::py_read_gt3x(path)
df = pygt3x::impute_zeros(res$data, res$dates, res$header)
out = wt_hees_2013(df)

\dontrun{

url = "https://github.com/THLfi/read.gt3x/files/3522749/GT3X\%2B.01.day.gt3x.zip"
destfile = tempfile(fileext = ".zip")
dl = download.file(url, destfile = destfile)
gt3x_file = unzip(destfile, exdir = tempdir())
gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
path = gt3x_file

res = pygt3x::py_read_gt3x(path)
df = pygt3x::impute_zeros(res$data, res$dates, res$header)
out = wt_hees_2013(df)
df = resample_acc(df)
out_100 = wt_hees_2013(df)
}
}
