% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_req.R
\name{wt_install_packages}
\alias{wt_install_packages}
\title{Install Packages Needed for \code{weartime}}
\usage{
wt_install_packages(
  packages = c("numpy", "pandas", "joblib", "psutil", "glob2", "tensorflow==2.0.0",
    "bitstring", "scipy", "resampy", "csv", "logging", "time", "datetime"),
  ...
)
}
\arguments{
\item{packages}{list of packages, passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{...}{additional arguments passed to passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}
}
\value{
Output from \code{\link[reticulate:py_install]{reticulate::py_install()}}
}
\description{
Install Packages Needed for \code{weartime}
}
