% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_req.R
\name{wt_packages_installed}
\alias{wt_packages_installed}
\title{Python modules installed}
\usage{
wt_packages_installed()
}
\value{
A logical
}
\description{
Python modules installed
}
\examples{
wt_packages_installed()
}
